try to solve another race condition that lost oauth tokens
The problem here was that SearchAll() takes a copy of a UserToken,
uses it repeatedly, changes its modification time, and puts it back
in storage.  But if the token expires in the middle of this process,
it changes underneath you, and is saved by the NonBrokenTokenStore,
and then the rest of the SearchAll() efforts fail, and then it stores
the now-obsolete token, overwriting the new one that was good.

Generally it is clear that I can't be trusted not to do unsafe things
with the UserToken object.  The fact that the oauth library might modify
it on any use makes any kind of read-modify-write a problem.  So this
turned into a major rewrite that makes the token set inaccessible except
through UserTokenSet methods.